Update type declarations

+++ includes/model/Listings.php
@@ -334,7 +334,7 @@ class Directorist_Listings {
 		$this->tags                     = empty( $this->params['tag'] ) ? '' : explode( ',', $this->params['tag'] );
 		$this->locations                = empty( $this->params['location'] ) ? '' : explode( ',', $this->params['location'] );
 		$this->ids                      = empty( $this->params['ids'] ) ? '' : explode( ',', $this->params['ids'] );
-		$this->columns                  = (int) atbdp_calculate_column( $this->params['columns'] );
+		$this->columns                  = atbdp_calculate_column( $this->params['columns'] );
 		$this->featured_only            = $this->params['featured_only'];
 		$this->popular_only             = $this->params['popular_only'];
 		$this->display_preview_image    = $this->params['display_preview_image'] == 'yes';
